+++ lib/sinatra/browse/parameter_types/any.rb
@@ -1,0 +1,13 @@
+# -*- coding: utf-8 -*-
+
+module Sinatra::Browse
+  module ParameterTypes
+
+    class Any < ParameterType
+      def coerce(value)
+        value
+      end
+    end
+
+  end
+end