+++ CHANGELOG.md
@@ -7,6 +7,10 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 * `Added` It is now possible to use sinatra-browse with named parameters in routes patterns.
 
+* `Added` New parameter types: `Hash`, `Array`, `Any`.
+
+* `Added` Error handling to type coercion. Parameters that fail to be converted to the correct type will return a 400 error with an error hash.
+
 * `Changed` Sinatra-browse now does its work in a condition instead of a before block. This was done to become able to use named parameters in route patterns but also means user defind conditions will be executed before sinatra-browse.
 
 ## [0.6.1] - 2015-07-07
@@ -44,8 +48,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ## [0.4] - 2014-05-29
 
 * `Added` Implemented two new String validations.
-    - min_length
-    - max_length
+    - min\_length
+    - max\_length
 
 ## [0.3] - 2014-05-08